refactor: Actions.ts reworked
- Removed namespaces
- Fixed imports
- @TODO: add documentation and move functionality to separate files if
  needed
